LayoutTests/media/media-source/media-managedmse-seek-and-evictable.html

<!DOCTYPE html>
<html>
<head>
    <title>mock-media-source</title>
    <script src="mock-media-source.js"></script>
    <script src="../video-test.js"></script>
    <script>
    var source;
    var sourceBuffer;
    var initSegment;
    var exception;
    var sampleSize = 72;

    function bufferedRanges() {
        var bufferedRanges = '[ ';
        var timeRanges = sourceBuffer.buffered;
        for (var i = 0 ; i < timeRanges.length ; i++) {
            if (i)
                bufferedRanges += ', ';
            bufferedRanges += timeRanges.start(i) + '...' + timeRanges.end(i);
        }
        bufferedRanges += ' ]';
        return bufferedRanges;
    }

    async function appendPtsRange(firstPts, lastPts) {
        var resultException = null;
        for (var pts = firstPts; pts <= lastPts; pts++) {
            try {
                consoleWrite('Appending PTS='+pts);
                sourceBuffer.appendBuffer(makeASample(pts, pts, 1, 1, 1, SAMPLE_FLAG.SYNC, 1));
                await waitFor(sourceBuffer, 'updateend');
            } catch (e) {
                resultException = e;
                sourceBuffer.abort();
                break;
            }
        }
        return resultException;
    }

    if (window.internals)
        internals.initializeMockMediaSource();

    window.addEventListener('load', async() => {
        findMediaElement();
        source = new ManagedMediaSource();

        const videoSource = document.createElement('source');
        videoSource.type = 'video/mock; codecs=mock';
        videoSource.src = URL.createObjectURL(source);
        run('video.disableRemotePlayback = true');
        video.appendChild(videoSource);

        await waitFor(source, 'sourceopen');
        sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock");
        initSegment = makeAInit(300, [makeATrack(1, 'mock', TRACK_KIND.VIDEO)]);
        sourceBuffer.appendBuffer(initSegment);
        await waitFor(sourceBuffer, 'updateend');

        testExpected('source.duration', 300, '==');
        testExpected('video.duration', 300, '==');

        // This allows bufering up to 100
        await internals.setMaximumSourceBufferSize(sourceBuffer, 100 * sampleSize);

        exception = await appendPtsRange(0, 99);
        // No past data, and all contiguous, nothing evictable and nothing evicted.
        testExpected('exception', null, '==');
        testExpected('internals.evictableSize(sourceBuffer)', 0, '==');
        testExpected('bufferedRanges()', '[ 0...100 ]', '==');
        
        // Attempting to append one more sample will fail.
        exception = await appendPtsRange(101, 180);
        testExpected('exception', 'QuotaExceededError: The quota has been exceeded.', '==');
        testExpected('internals.evictableSize(sourceBuffer)', 0, '==');

        exception = await appendPtsRange(200, 250);
        testExpected('exception', 'QuotaExceededError: The quota has been exceeded.', '==');
        testExpected('internals.evictableSize(sourceBuffer)', 0, '==');

        run('video.currentTime = 175');
        testExpected('video.currentTime', 175, '==');

        // Samples before currentTime are evictable.
        await Promise.all([ waitFor(video, 'seeked', true), appendPtsRange(175, 200) ]);
        testExpected('bufferedRanges()', '[ 30...100, 175...201 ]', '==');
        testExpected('internals.evictableSize(sourceBuffer)', (100 - 30) * sampleSize, '==');
       
        sourceBuffer.remove(0, 500);
        await waitFor(sourceBuffer, 'updateend');
        
        run('video.currentTime = 0');
        testExpected('video.currentTime', 0, '==');

        await Promise.all([ waitFor(video, 'seeked', true), appendPtsRange(0, 49) ]);
        testExpected('bufferedRanges()', '[ 0...50 ]', '==');
        testExpected('internals.evictableSize(sourceBuffer)', 0, '==');

        exception = await appendPtsRange(200, 249);
        testExpected('exception', null, '==');
        testExpected('bufferedRanges()', '[ 0...50, 200...250 ]', '==');
        testExpected('internals.evictableSize(sourceBuffer)', (250 - 200) * sampleSize, '==');

        run('video.currentTime = 60');
        testExpected('video.currentTime', 60, '==');

        // Appending new data, will evict past data up to currentTime - 3s.
        await Promise.all([ waitFor(video, 'seeked', true), appendPtsRange(55, 99) ]);
        testExpected('bufferedRanges()', '[ 45...50, 55...100, 200...250 ]', '==');
        testExpected('internals.evictableSize(sourceBuffer)', ((50 - 45) + (video.currentTime - 3 - 55) + (250 - 200)) * sampleSize, '==');

        // Appending new data, will evict future data, no contiguous with the currently playing range.
        run('video.currentTime = 0');
        testExpected('video.currentTime', 0, '==');
        await Promise.all([ waitFor(video, 'seeked', true), appendPtsRange(0, 49) ]);
        testExpected('bufferedRanges()', '[ 0...50, 55...100 ]', '==');
        testExpected('internals.evictableSize(sourceBuffer)', (100 - 55) * 72, '==');

        endTest();
     });
    </script>
</head>
<body>
    <video></video>
</body>
</html>
